Extract resolvers and mutations for shipping
